adicionado acl

+++ app/Http/Controllers/API/UserController.php
@@ -19,7 +19,8 @@ class UserController extends Controller
     }
     public function index()
     {
-        return User::latest()->paginate(10);
+        $this->authorize("isAdmin");
+        return User::latest()->paginate(20);
     }
 
     /**
@@ -51,9 +52,37 @@ class UserController extends Controller
     public function updateProfile(Request $request)
     {
         $user = auth('api')->user();
+      
+        $request->validate([
+            'name'      =>    'required|string|max:191',
+            'email'     =>    'required|string|email|max:191|unique:users,email,'.$user->id,
+            'password'  =>    'sometimes|required|min:6'
+        ]);
 
-        return $request->photo;
-        // return ['message'=>'Success'];
+
+        $currentPhoto = $user->photo;
+        /**if the photo in database is diferent fron this phtoto, then post */
+        if($request->photo != $currentPhoto){
+            $name = time().'.' . explode('/', explode(':',substr($request->photo,0,strpos($request->photo,';')))[1])[1];
+            
+            \Image::make($request->photo)->save(public_path('img/profile/').$name); 
+
+            $request->merge(['photo' => $name ]);
+
+            $user_photo = public_path('img/profile/').$currentPhoto;
+
+            if( file_exists( $user_photo )){
+                @unlink($user_photo);
+            }
+        }
+
+        if(!empty($request->password)){
+            $request->merge(['password' => bcrypt($request->password)]);
+          
+        }
+        $user->update($request->all());
+        
+        return ['message'=>'Success'];
     }
     public function profile()
     {
@@ -89,6 +118,10 @@ class UserController extends Controller
             'password'  =>    'sometimes|min:6'
         ]);
 
+        if(!empty($request->password)){
+            $request->merge(['password' => bcrypt($request->password)]);
+          
+        }
 
         $user->update($request->all());
 
@@ -105,11 +138,20 @@ class UserController extends Controller
     public function destroy($id)
     {
         //
+
+        $this->authorize("isAdmin");
         $user= User::findOrFail($id);
         
         
         $user->delete();
         
         return ['message'=>'User Deleted'];
+    }
+    private function validation(){
+        return [
+            'name'      =>    'required|string|max:191',
+            'email'     =>    'required|string|email|max:191|unique:users',
+            'password'  =>    'required|string|min:6'
+        ];
     }
 }
